AI股票分析 - 自用/requirements.txt
akshare>=1.14
pandas>=2.2
requests>=2.32
python-dotenv>=1.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
mcp>=1.0.0
uvicorn>=0.34.0
# Python >= 3.10 required
